+++ packages/create-mastra/CHANGELOG.md
@@ -1,5 +1,9 @@
 # create-mastra
 
+## 1.10.1-alpha.7
+
+## 1.10.1-alpha.6
+
 ## 1.10.1-alpha.5
 
 ## 1.10.1-alpha.4